refactor: 메세지 입력창 messenger 폴더로 이동
